+++ src/asteroid.hpp
@@ -12,13 +12,13 @@ class Asteroid
 public:
   Vector2 position{};
   Vector2 velocity{};
-  int size{ 2 };
-  int max_life { 1 };
-  int life{ max_life };
+  uint8_t size{ 2 };
+  uint8_t max_life { 1 };
+  uint8_t life{ max_life };
   mutable Sprite sprite{ "resources/asteroid.aseprite" };
   Mask mask{};
 
-  [[nodiscard]] static Asteroid create(const Vector2 &position, int size);
+  [[nodiscard]] static Asteroid create(const Vector2 &position, uint8_t size);
 
   bool update();
   void die();
